Refactored Presenters, Tests, URLs, and Views
 All of these files were growing quite large. This refactoring splits them by major component--enrollment, engagement, performance. The resulting files are smaller and more maintainable.
