+++ analytics_dashboard/analytics_dashboard/settings/test.py
@@ -21,3 +21,7 @@ DATABASES = {
 
 ENABLE_AUTO_AUTH = True
 SOCIAL_AUTH_EDX_OIDC_URL_ROOT = 'http://example.com'
+
+COURSE_API_URL = 'http://course-api-host'
+COURSE_API_VERSION = 'v0'
+COURSE_API_KEY = 'edx'
