+++ analytics_dashboard/courses/presenters/engagement.py
@@ -1,0 +1,86 @@
+import datetime
+
+from waffle import switch_is_active
+from analyticsclient.client import Client
+import analyticsclient.constants.activity_type as AT
+
+from courses.presenters import BasePresenter
+
+
+class CourseEngagementPresenter(BasePresenter):
+    """
+    Presenter for the engagement page.
+    """
+
+    def get_activity_types(self):
+        activities = [AT.ANY, AT.PLAYED_VIDEO, AT.ATTEMPTED_PROBLEM]
+
+        # Include forum activity only if feature is enabled.
+        if switch_is_active('show_engagement_forum_activity'):
+            activities.append(AT.POSTED_FORUM)
+
+        return activities
+
+    def _build_trend_week(self, trend_types, week_ending, api_trend):
+        trend_week = {'weekEnding': week_ending.isoformat()}
+        for trend_type in trend_types:
+            if trend_type in api_trend:
+                trend_week[trend_type] = api_trend[trend_type] or 0
+            else:
+                trend_week[trend_type] = 0
+        return trend_week
+
+    def _build_trend(self, api_trends):
+        """
+        Format activity trends for specified date range and return results with
+        zeros filled in for all activities.
+        """
+        trend_types = self.get_activity_types()
+        trends = []
+
+        # add zeros for the week prior if we only have a single point (prevents just a single point in the chart)
+        if len(api_trends) == 1:
+            interval_end = self.parse_api_datetime(api_trends[0]['interval_end'])
+            week_ending = interval_end.date() - datetime.timedelta(days=8)
+            trends.append(self._build_trend_week(trend_types, week_ending, {}))
+
+        # fill in gaps in activity with zero for display (api doesn't return
+        # the field if no data exists for it, so we fill in the zeros here)
+        for datum in api_trends:
+            # convert end of interval to ending day of week
+            interval_end = self.parse_api_datetime(datum['interval_end'])
+            week_ending = interval_end.date() - datetime.timedelta(days=1)
+            trends.append(self._build_trend_week(trend_types, week_ending, datum))
+
+        return trends
+
+    def _build_summary(self, api_trends):
+        """
+        Format all summary numbers and week ending time.
+        """
+
+        most_recent = api_trends[-1]
+        summary = {
+            'last_updated': self.parse_api_datetime(most_recent['created'])
+        }
+
+        # Fill in gaps in the summary if no data found so we can display a proper message
+        activity_types = self.get_activity_types()
+        for activity_type in activity_types:
+            if activity_type in most_recent:
+                summary[activity_type] = most_recent[activity_type]
+            else:
+                summary[activity_type] = None
+
+        return summary
+
+    def get_summary_and_trend_data(self):
+        """
+        Retrieve recent summary and all historical trend data.
+        """
+        # setting the end date (exclusive) to the next day retrieves data as soon as it's available
+        end_date = (datetime.datetime.utcnow() + datetime.timedelta(days=1)).strftime(Client.DATE_FORMAT)
+        api_trends = self.course.activity(start_date=None, end_date=end_date)
+        summary = self._build_summary(api_trends)
+        trends = self._build_trend(api_trends)
+        return summary, trends
