+++ analytics_dashboard/courses/presenters/enrollment.py
@@ -2,20 +2,15 @@ import copy
 import datetime
 import logging
 
-from collections import namedtuple
-
 from django.utils.translation import ugettext_lazy as _
-from django.conf import settings
 from django_countries import countries
 from waffle import switch_is_active
-from analyticsclient.client import Client
-import analyticsclient.constants.activity_type as AT
+from analyticsclient.constants import demographic, UNKNOWN_COUNTRY_CODE, enrollment_modes
 import analyticsclient.constants.education_level as EDUCATION_LEVEL
 import analyticsclient.constants.gender as GENDER
-from analyticsclient.constants import demographic, UNKNOWN_COUNTRY_CODE, enrollment_modes
-from analyticsclient.exceptions import NotFoundError
 
 import courses.utils as utils
+from courses.presenters import BasePresenter
 
 
 logger = logging.getLogger(__name__)
@@ -80,261 +75,6 @@ EDUCATION_ORDER = {
     EDUCATION_LEVEL.DOCTORATE: 7,
     EDUCATION_LEVEL.OTHER: 8,
 }
-
-
-class BasePresenter(object):
-    """
-    This is the base class for the pages and sets up the analytics client
-    for the presenters to use to access the data API.
-    """
-
-    def __init__(self, course_id, timeout=5):
-        self.client = Client(base_url=settings.DATA_API_URL,
-                             auth_token=settings.DATA_API_AUTH_TOKEN,
-                             timeout=timeout)
-        self.course_id = course_id
-        self.course = self.client.courses(self.course_id)
-
-    def get_current_date(self):
-        return datetime.datetime.utcnow().strftime(Client.DATE_FORMAT)
-
-    @staticmethod
-    def parse_api_date(s):
-        """ Parse a string according to the API date format. """
-        return datetime.datetime.strptime(s, Client.DATE_FORMAT).date()
-
-    @staticmethod
-    def parse_api_datetime(s):
-        """ Parse a string according to the API datetime format. """
-        return datetime.datetime.strptime(s, Client.DATETIME_FORMAT)
-
-    @staticmethod
-    def strip_time(s):
-        return s[:-7]
-
-    @staticmethod
-    def sum_counts(data):
-        return sum(datum['count'] for datum in data)
-
-
-# stores the answer distribution return from CoursePerformancePresenter
-AnswerDistributionEntry = namedtuple('AnswerDistributionEntry', [
-    'last_updated',
-    'questions',
-    'active_question',
-    'answer_distribution',
-    'answer_distribution_limited',
-    'is_random',
-    'answer_type',
-    'problem_part_description'
-])
-
-
-class CoursePerformancePresenter(BasePresenter):
-    """
-    Presenter for the performance page.
-    """
-
-    # limit for the number of bars to display in the answer distribution chart
-    CHART_LIMIT = 12
-
-    def get_answer_distribution(self, problem_id, problem_part_id):
-        """
-        Retrieve answer distributions for a particular module/problem and problem part.
-        """
-
-        module = self.client.modules(self.course_id, problem_id)
-
-        api_response = module.answer_distribution()
-        questions = self._build_questions(api_response)
-
-        filtered_active_question = [i for i in questions if i['part_id'] == problem_part_id]
-        if len(filtered_active_question) == 0:
-            raise NotFoundError
-        else:
-            active_question = filtered_active_question[0]['question']
-
-        answer_distributions = self._build_answer_distribution(api_response, problem_part_id)
-        problem_part_description = self._build_problem_description(problem_part_id, questions)
-
-        is_random = self._is_answer_distribution_random(answer_distributions)
-        answer_distribution_limited = None
-        if not is_random:
-            # only display the top in the chart
-            answer_distribution_limited = answer_distributions[:self.CHART_LIMIT]
-
-        answer_type = self._get_answer_type(answer_distributions)
-        last_updated = self.parse_api_datetime(answer_distributions[0]['created'])
-
-        return AnswerDistributionEntry(last_updated, questions, active_question, answer_distributions,
-                                       answer_distribution_limited, is_random, answer_type, problem_part_description)
-
-    def _build_problem_description(self, problem_part_id, questions):
-        """ Returns the displayable problem name. """
-        problem = [q for q in questions if q['part_id'] == problem_part_id][0]
-        if problem['problem_name']:
-            return u'{0} - {1}'.format(problem['problem_name'], problem['question'])
-        return problem['question']
-
-    def _get_answer_type(self, answer_distributions):
-        """ Returns either answer_value_text or answer_value_numeric. """
-
-        # answer field for numeric values returned from API
-        numeric_field = 'answer_value_numeric'
-        # answer field for text values returned from API
-        text_field = 'answer_value_text'
-
-        # returns the first field found to be filled
-        for ad in answer_distributions:
-            if ad[numeric_field] is not None:
-                return numeric_field
-            elif ad[text_field] is not None:
-                return text_field
-
-        # default to text answer if both fields were null/none
-        return text_field
-
-    def _is_answer_distribution_random(self, answer_distributions):
-        """
-        Problems are considered randomized if variant is populated with values
-        greater than 1.
-        """
-        for ad in answer_distributions:
-            variant = ad['variant']
-            if variant is not None and variant is not 1:
-                return True
-        return False
-
-    def _build_questions(self, answer_distributions):
-        """
-        Builds the questions and part_id from the answer distribution. Displayed
-        drop down.
-        """
-        questions = []
-        part_id_to_problem = {}
-
-        # Collect unique questions from the answer distribution
-        for question_answer in answer_distributions:
-            question = question_answer.get('question_text', None)
-            problem_name = question_answer.get('problem_display_name', None)
-            part_id_to_problem[question_answer['part_id']] = {
-                'question': question,
-                'problem_name': problem_name
-            }
-
-        for part_id, problem in part_id_to_problem.iteritems():
-            questions.append({
-                'part_id': part_id,
-                'question': problem['question'],
-                'problem_name': problem['problem_name']
-            })
-
-        utils.sorting.natural_sort(questions, 'part_id')
-
-        # add an enumerated label
-        has_parts = len(questions) > 1
-        for i, question in enumerate(questions):
-            text = question['question']
-            question_num = i + 1
-            template = _('Submissions')
-            if text:
-                if has_parts:
-                    template = _('Submissions for Part {part_number}: {part_description}')
-                else:
-                    template = _('Submissions: {part_description}')
-            else:
-                if has_parts:
-                    template = _('Submissions for Part {part_number}')
-
-            # pylint: disable=no-member
-            question['question'] = template.format(part_number=question_num, part_description=text)
-
-        return questions
-
-    def _build_answer_distribution(self, api_response, problem_part_id):
-        """ Filter for this problem part and sort descending order. """
-        answer_distributions = [i for i in api_response if i['part_id'] == problem_part_id]
-        answer_distributions = sorted(answer_distributions, key=lambda a: -a['count'])
-        return answer_distributions
-
-
-class CourseEngagementPresenter(BasePresenter):
-    """
-    Presenter for the engagement page.
-    """
-
-    def get_activity_types(self):
-        activities = [AT.ANY, AT.PLAYED_VIDEO, AT.ATTEMPTED_PROBLEM]
-
-        # Include forum activity only if feature is enabled.
-        if switch_is_active('show_engagement_forum_activity'):
-            activities.append(AT.POSTED_FORUM)
-
-        return activities
-
-    def _build_trend_week(self, trend_types, week_ending, api_trend):
-        trend_week = {'weekEnding': week_ending.isoformat()}
-        for trend_type in trend_types:
-            if trend_type in api_trend:
-                trend_week[trend_type] = api_trend[trend_type] or 0
-            else:
-                trend_week[trend_type] = 0
-        return trend_week
-
-    def _build_trend(self, api_trends):
-        """
-        Format activity trends for specified date range and return results with
-        zeros filled in for all activities.
-        """
-        trend_types = self.get_activity_types()
-        trends = []
-
-        # add zeros for the week prior if we only have a single point (prevents just a single point in the chart)
-        if len(api_trends) == 1:
-            interval_end = self.parse_api_datetime(api_trends[0]['interval_end'])
-            week_ending = interval_end.date() - datetime.timedelta(days=8)
-            trends.append(self._build_trend_week(trend_types, week_ending, {}))
-
-        # fill in gaps in activity with zero for display (api doesn't return
-        # the field if no data exists for it, so we fill in the zeros here)
-        for datum in api_trends:
-            # convert end of interval to ending day of week
-            interval_end = self.parse_api_datetime(datum['interval_end'])
-            week_ending = interval_end.date() - datetime.timedelta(days=1)
-            trends.append(self._build_trend_week(trend_types, week_ending, datum))
-
-        return trends
-
-    def _build_summary(self, api_trends):
-        """
-        Format all summary numbers and week ending time.
-        """
-
-        most_recent = api_trends[-1]
-        summary = {
-            'last_updated': self.parse_api_datetime(most_recent['created'])
-        }
-
-        # Fill in gaps in the summary if no data found so we can display a proper message
-        activity_types = self.get_activity_types()
-        for activity_type in activity_types:
-            if activity_type in most_recent:
-                summary[activity_type] = most_recent[activity_type]
-            else:
-                summary[activity_type] = None
-
-        return summary
-
-    def get_summary_and_trend_data(self):
-        """
-        Retrieve recent summary and all historical trend data.
-        """
-        # setting the end date (exclusive) to the next day retrieves data as soon as it's available
-        end_date = (datetime.datetime.utcnow() + datetime.timedelta(days=1)).strftime(Client.DATE_FORMAT)
-        api_trends = self.course.activity(start_date=None, end_date=end_date)
-        summary = self._build_summary(api_trends)
-        trends = self._build_trend(api_trends)
-        return summary, trends
 
 
 class CourseEnrollmentPresenter(BasePresenter):
