+++ analytics_dashboard/courses/presenters/performance.py
@@ -1,0 +1,148 @@
+from collections import namedtuple
+
+from django.utils.translation import ugettext_lazy as _
+from analyticsclient.exceptions import NotFoundError
+
+import courses.utils as utils
+from courses.presenters import BasePresenter
+
+
+# Stores the answer distribution return from CoursePerformancePresenter
+AnswerDistributionEntry = namedtuple('AnswerDistributionEntry', [
+    'last_updated',
+    'questions',
+    'active_question',
+    'answer_distribution',
+    'answer_distribution_limited',
+    'is_random',
+    'answer_type',
+    'problem_part_description'
+])
+
+
+class CoursePerformancePresenter(BasePresenter):
+    """
+    Presenter for the performance page.
+    """
+
+    # limit for the number of bars to display in the answer distribution chart
+    CHART_LIMIT = 12
+
+    def get_answer_distribution(self, problem_id, problem_part_id):
+        """
+        Retrieve answer distributions for a particular module/problem and problem part.
+        """
+
+        module = self.client.modules(self.course_id, problem_id)
+
+        api_response = module.answer_distribution()
+        questions = self._build_questions(api_response)
+
+        filtered_active_question = [i for i in questions if i['part_id'] == problem_part_id]
+        if len(filtered_active_question) == 0:
+            raise NotFoundError
+        else:
+            active_question = filtered_active_question[0]['question']
+
+        answer_distributions = self._build_answer_distribution(api_response, problem_part_id)
+        problem_part_description = self._build_problem_description(problem_part_id, questions)
+
+        is_random = self._is_answer_distribution_random(answer_distributions)
+        answer_distribution_limited = None
+        if not is_random:
+            # only display the top in the chart
+            answer_distribution_limited = answer_distributions[:self.CHART_LIMIT]
+
+        answer_type = self._get_answer_type(answer_distributions)
+        last_updated = self.parse_api_datetime(answer_distributions[0]['created'])
+
+        return AnswerDistributionEntry(last_updated, questions, active_question, answer_distributions,
+                                       answer_distribution_limited, is_random, answer_type, problem_part_description)
+
+    def _build_problem_description(self, problem_part_id, questions):
+        """ Returns the displayable problem name. """
+        problem = [q for q in questions if q['part_id'] == problem_part_id][0]
+        if problem['problem_name']:
+            return u'{0} - {1}'.format(problem['problem_name'], problem['question'])
+        return problem['question']
+
+    def _get_answer_type(self, answer_distributions):
+        """ Returns either answer_value_text or answer_value_numeric. """
+
+        # answer field for numeric values returned from API
+        numeric_field = 'answer_value_numeric'
+        # answer field for text values returned from API
+        text_field = 'answer_value_text'
+
+        # returns the first field found to be filled
+        for ad in answer_distributions:
+            if ad[numeric_field] is not None:
+                return numeric_field
+            elif ad[text_field] is not None:
+                return text_field
+
+        # default to text answer if both fields were null/none
+        return text_field
+
+    def _is_answer_distribution_random(self, answer_distributions):
+        """
+        Problems are considered randomized if variant is populated with values
+        greater than 1.
+        """
+        for ad in answer_distributions:
+            variant = ad['variant']
+            if variant is not None and variant is not 1:
+                return True
+        return False
+
+    def _build_questions(self, answer_distributions):
+        """
+        Builds the questions and part_id from the answer distribution. Displayed
+        drop down.
+        """
+        questions = []
+        part_id_to_problem = {}
+
+        # Collect unique questions from the answer distribution
+        for question_answer in answer_distributions:
+            question = question_answer.get('question_text', None)
+            problem_name = question_answer.get('problem_display_name', None)
+            part_id_to_problem[question_answer['part_id']] = {
+                'question': question,
+                'problem_name': problem_name
+            }
+
+        for part_id, problem in part_id_to_problem.iteritems():
+            questions.append({
+                'part_id': part_id,
+                'question': problem['question'],
+                'problem_name': problem['problem_name']
+            })
+
+        utils.sorting.natural_sort(questions, 'part_id')
+
+        # add an enumerated label
+        has_parts = len(questions) > 1
+        for i, question in enumerate(questions):
+            text = question['question']
+            question_num = i + 1
+            template = _('Submissions')
+            if text:
+                if has_parts:
+                    template = _('Submissions for Part {part_number}: {part_description}')
+                else:
+                    template = _('Submissions: {part_description}')
+            else:
+                if has_parts:
+                    template = _('Submissions for Part {part_number}')
+
+            # pylint: disable=no-member
+            question['question'] = template.format(part_number=question_num, part_description=text)
+
+        return questions
+
+    def _build_answer_distribution(self, api_response, problem_part_id):
+        """ Filter for this problem part and sort descending order. """
+        answer_distributions = [i for i in api_response if i['part_id'] == problem_part_id]
+        answer_distributions = sorted(answer_distributions, key=lambda a: -a['count'])
+        return answer_distributions
