+++ analytics_dashboard/courses/templates/courses/enrollment_demographics_education.html
@@ -90,7 +90,7 @@
 
         <div class="section-actions">
             {% if js_data.course.education %}
-                <a href="{% url 'courses:csv_enrollment_demographics_education' course_id=course_id %}"
+                <a href="{% url 'courses:csv:enrollment_demographics_education' course_id=course_id %}"
                    class="btn btn-default"
                    data-role="enrollment-csv" data-track-type="click"
                    data-track-event="edx.bi.csv.downloaded"
