+++ analytics_dashboard/courses/templates/courses/enrollment_demographics_gender.html
@@ -55,7 +55,7 @@
 
         <div class="section-actions">
             {% if js_data.course.genderTrend %}
-                <a href="{% url 'courses:csv_enrollment_demographics_gender' course_id=course_id %}"
+                <a href="{% url 'courses:csv:enrollment_demographics_gender' course_id=course_id %}"
                    class="btn btn-default"
                    data-role="enrollment-csv" data-track-type="click"
                    data-track-event="edx.bi.csv.downloaded"
