+++ analytics_dashboard/courses/templates/courses/performance_answer_distribution.html
@@ -59,7 +59,7 @@
                                     {% for question in questions %}
                                         <li role="presentation">
                                             <a role="menuitem"tabindex="-1" class="truncate"
-                                               href="{% url 'courses:performance_answerdistribution' course_id=course_id content_id=problem_id problem_part_id=question.part_id %}">
+                                               href="{% url 'courses:performance:answer_distribution' course_id=course_id content_id=problem_id problem_part_id=question.part_id %}">
                                                 {{question.question}}
                                             </a>
                                         </li>
@@ -111,7 +111,7 @@
         <h4 class="section-title">{% trans "Submission Counts" %}</h4>
         <div class="section-actions">
             {% if js_data.course.answerDistribution %}
-                <a href="{% url 'courses:csv_performance_answerdistribution' course_id=course_id content_id=problem_id problem_part_id=problem_part_id %}"
+                <a href="{% url 'courses:csv:performance_answer_distribution' course_id=course_id content_id=problem_id problem_part_id=problem_part_id %}"
                    class="btn btn-default"
                    data-role="performance-csv" data-track-type="click"
                    data-track-event="edx.bi.csv.downloaded"
